Remove location obfuscation field for now.
It seems incorrect to me that the stream encoder should receive the
precise location from the wallet, and then be responsible for
obfuscating that location; rather I think that if the user wishes to
obfuscate their location they should do that before adding their device
into the DECODE world. Maybe this should be in the wallet or perhaps
even the onboarding application.